pyproject.toml
[project]
name = "dab_project"
version = "0.0.1"
authors = [{ name = "yusuf.kamorudeen@devoteam.com" }]
requires-python = ">=3.10,<3.13"
dependencies = [
    # Any dependencies for jobs and pipelines in this project can be added here
    # See also https://docs.databricks.com/dev-tools/bundles/library-dependencies
    #
    # LIMITATION: for pipelines, dependencies are cached during development;
    # add dependencies to the 'environment' section of your pipeline.yml file instead
    "pytest>=9.0.2",
]

[dependency-groups]
dev = [
    "pytest",
    "databricks-dlt",
]

[project.scripts]
main = "dab_project.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/citibike", "src/dab_project","src/utils"]

[tool.black]
line-length = 125
